Restructured models directory to make it a module
